% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODPSCP-package.R
\docType{package}
\name{ODPSCP}
\alias{ODPSCP-package}
\alias{ODPSCP}
\title{ODPSCP}
\description{
The ODPSCP shiny interface comes in a series of UI and server plus helper functions
that support the standardized reporting of prioritization results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iiasa/ODPSCP}
  \item Report bugs at \url{https://github.com/iiasa/ODPSCP/issues}
}

}
\author{
\strong{Maintainer}: Martin Jung \email{jung@iiasa.ac.at}

}
\keyword{internal}
